+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 09-NOV-2025
-- update detail page to fully using compose 
+- update main page & detail page to fully using compose
+
+05-NOV-2025
+- migrate to lib toml gradle kts version